+++ src/gamelogic/base/src/cgame/cg_main.c
@@ -90,7 +90,7 @@ intptr_t vmMain( intptr_t command, intptr_t arg0, intptr_t arg1, intptr_t arg2, 
 	  return 0;
 
     default:
-      CG_Error( "vmMain: unknown command %i", command );
+      CG_Error( "vmMain: unknown command %i", (int)command );
       break;
   }
